[Move #247] store이름을 kebab case로 변경
